requirements.txt
line-bot-sdk
flask
requests
bs4
# selenium==3.8.0
# APScheduler
# paho-mqtt
fake_useragent
# openai
python-dotenv